Restructured codebase for namespace pip install
The codebase has been restructured to be easily importable after
installation by pip. We have created a new directory/ namespace called
ezdb which will hold all our database abstractions in future.
Still have no been able to solve the docs exclusion problem, but that is
next on the chop.
